Adapt for upcoming release of new INSPIRE version
We will release the new INSPIRE version (currently at https://labs.inspirehep.net) as the main website on Monday. The old website will remain available at https://old.inspirehep.net for a couple of months. The new API is not 100% ready for public use, so this commit simply switches to the URL of the old website instead of using the new one.