Avoid creating duplicated bibtex entries from iNSPIRE
iNSPIRE sometimes returns bibtex entries with a different key from the
one used to query the database. For example, searching for
"Monitor:2017mdv" returns a valid bibtex entry, but with key
"LIGOScientific:2017zic". This caused fillbib and bibtex to be confused.
This fix replaces the iNSPIRE key in the bibtex with the key used for
the search, so that the entry is correctly found by both bibtex and
fillbib.